# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "matcheck Mobile API",
    "version": "1.0.0",
    "description": "API для внешних клиентов matcheck (мобильные приложения, интеграции).\n\nБазовый URL: `https://matcheck.fvds.ru/api/v1/`.\n\nAuth: Bearer JWT (Ed25519). Для мобильных клиентов используется заголовок `X-Client-Type: mobile` — в этом случае refresh-token возвращается в теле ответа `/auth/login` и `/auth/refresh`, cookies не устанавливаются.\n\nПодробная документация: см. `docs/MOBILE_API.md` в репозитории.",
    "contact": {
      "name": "matcheck",
      "url": "https://matcheck.fvds.ru"
    }
  },
  "servers": [
    {
      "url": "https://matcheck.fvds.ru",
      "description": "Production"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ed25519 JWT access-token, полученный из POST /auth/login или /auth/refresh. TTL 900 сек, обновляется через /auth/refresh."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error"
        ]
      },
...
      "PhotoPresignRequest": {
        "type": "object",
        "properties": {
          "operationKind": {
            "type": "string",
            "enum": [
              "delivery",
              "shipment"
            ],
            "default": "delivery"
          },
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "deliveryId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "document",
              "cargo",
              "vehicle",
              "other"
            ]
          },
          "contentHash": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "idempotencyKey": {
            "type": "string",
            "format": "uuid"
          },
          "contentType": {
            "type": "string",
            "default": "image/jpeg"
          },
          "thumbContentHash": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          }
        },
        "required": [
          "kind",
          "contentHash",
          "idempotencyKey"
        ]
      },
      "PhotoPresignResponse": {
        "type": "object",
        "properties": {
          "photoId": {
            "type": "string",
            "format": "uuid"
          },
          "s3Key": {
            "type": "string"
          },
          "thumbS3Key": {
            "type": [
              "string",
              "null"
            ]
          },
          "uploadUrl": {
            "type": "string"
          },
          "thumbUploadUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "expiresIn": {
            "type": "number"
          },
          "alreadyExists": {
            "type": "boolean"
          }
        },
        "required": [
          "photoId",
          "s3Key",
          "thumbS3Key",
          "uploadUrl",
          "thumbUploadUrl",
          "expiresIn",
          "alreadyExists"
        ]
      },
      "PhotoGetUrlResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number"
          }
        },
        "required": [
          "url",
          "expiresIn"
        ]
      },
      "PhotoDeleteResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
      "PhotoConfirmResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "uploadedAt": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "uploadedAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/photos/presign": {
      "post": {
        "tags": [
          "Photos"
        ],
        "summary": "Получить presigned PUT-URL для загрузки фото в S3",
        "description": "Двухэтапный pipeline: клиент сжимает фото, считает SHA-256 → presign → PUT в S3. Дедупликация по contentHash — если фото уже существует, возвращается alreadyExists=true и клиент пропускает PUT. TTL presigned URL: 300 сек.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoPresignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned URL выдан (или alreadyExists=true)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoPresignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизовано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/photos/{id}/confirm": {
      "post": {
        "tags": [
          "Photos"
        ],
        "summary": "Подтвердить успешную загрузку фото в S3",
        "description": "Клиент вызывает после успешного PUT в S3. Сервер делает S3.HEAD и, если объект существует, проставляет `uploaded_at = now()`. Это защищает запись от cleanup-job orphan-записей. Idempotent: повторный вызов возвращает прежний uploaded_at. Inspector_kpp может подтвердить только фото своего объекта (по siteId).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Подтверждено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoConfirmResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизовано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Фото не найдено или ещё не загружено в S3 (not_in_s3)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "S3 недоступна (s3_unavailable) — повторить позже",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/photos/{id}/url": {
      "get": {
        "tags": [
          "Photos"
        ],
        "summary": "Получить presigned GET-URL для просмотра фото",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Если true — возвращается URL для thumbnail"
            },
            "required": false,
            "name": "thumb",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Presigned URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoGetUrlResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизовано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Фото не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/photos/{id}": {
      "delete": {
        "tags": [
          "Photos"
        ],
        "summary": "Удалить фото (только admin)",
        "description": "Удаляет запись из delivery_photos и связанные объекты в S3 (основной + thumb). Доступно только пользователям с ролью admin. Ошибки S3 не валят запрос — логируются. Если родительский документ помечен на удаление — 409 `pending_deletion`.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Фото удалено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoDeleteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизовано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён (недостаточно прав)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Фото не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Родительский документ помечен на удаление",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}